+++ src/components/input/input.test.jsx
@@ -1,152 +1,123 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import { fireEvent, render, screen } from '@testing-library/react';
+import React, {
+  createRef,
+} from 'react';
+import {
+  fireEvent,
+  render,
+  screen,
+} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import Input from './input.jsx';
 
 describe('Input', () => {
+  const requiredProps = {
+    id: 'foo',
+    label: 'the label',
+  };
+
   it('has defaults', () => {
-    const tree = renderer
-      .create((
-        <Input
-          id="foo"
-          label="I am a label for accessibility"
-        />
-      )).toJSON();
-    expect(tree).toMatchSnapshot();
+    const ref = createRef();
+    render(<Input {...requiredProps} ref={ref} />);
+    expect(document.body).toMatchSnapshot();
+    expect(ref.current).toMatchSnapshot();
+  });
+
+  describe('autoFocus API', () => {
+    it('sets the autoFocus attribute', () => {
+      render(<Input {...requiredProps} autoFocus />);
+      expect(screen.getByLabelText('the label')).toHaveFocus();
+    });
   });
 
   describe('className API', () => {
     it('sets <input> className', () => {
-      render((
-        <Input
-          id="foo"
-          className="test-class"
-          label="the label"
-          invalid
-        />
-      ));
+      render(<Input {...requiredProps} className="test-class" />);
       expect(screen.getByLabelText('the label')).toHaveClass('test-class');
     });
   });
 
   describe('cssContainer API', () => {
     it('sets input container className', () => {
-      render((
-        <Input
-          id="foo"
-          cssContainer="test-class"
-          label="the label"
-          invalid
-        />
-      ));
+      render(<Input {...requiredProps} cssContainer="test-class" />);
       expect(screen.getByLabelText('the label').parentElement.parentElement).toHaveClass('test-class');
     });
   });
 
-  describe('defaultValue API', () => {
-    it('sets input value but input remains uncontrolled', () => {
-      render((
-        <Input
-          id="foo"
-          defaultValue="this is a test value"
-          label="the label"
-        />
-      ));
-
-      expect(screen.getByLabelText('the label')).toHaveValue('this is a test value');
-
-      userEvent.type(screen.getByLabelText('the label'), ' plus some more text');
-
-      expect(screen.getByLabelText('the label')).toHaveValue('this is a test value plus some more text');
+  describe('dirty API', () => {
+    it('is unset without any changes', () => {
+      const ref = createRef();
+      render(<Input {...requiredProps} ref={ref} />);
+      expect(ref.current.dirty).toBe(false);
     });
-  });
 
-  describe('disabled API', () => {
-    it('sets the disabled attribute', () => {
-      render((
-        <Input
-          id="foo"
-          disabled
-          label="the label"
-        />
-      ));
-      expect(screen.getByLabelText('the label')).toBeDisabled();
+    it('is set with any changes', () => {
+      const ref = createRef();
+      render(<Input {...requiredProps} ref={ref} />);
+      userEvent.type(screen.getByLabelText('the label'), 'a');
+      expect(ref.current.dirty).toBe(true);
+    });
+
+    it('is unset with changes back to initial state', () => {
+      const ref = createRef();
+      render(<Input {...requiredProps} ref={ref} />);
+      userEvent.type(screen.getByLabelText('the label'), 'a');
+      userEvent.type(screen.getByLabelText('the label'), '{backspace}');
+      expect(ref.current.dirty).toBe(false);
+    });
+
+    it('is unset with changes and new value prop', () => {
+      const ref = createRef();
+      const { rerender } = render(<Input {...requiredProps} ref={ref} />);
+      userEvent.type(screen.getByLabelText('the label'), 'a');
+      rerender(<Input {...requiredProps} ref={ref} value="a" />);
+      expect(ref.current.dirty).toBe(false);
     });
   });
 
   describe('id API', () => {
     it('sets the ID', () => {
-      render((
-        <Input
-          id="test-id"
-          label="the label"
-        />
-      ));
+      render(<Input {...requiredProps} id="test-id" />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('id', 'test-id');
+    });
+
+    it('is set on the ref', () => {
+      const ref = createRef();
+      render(<Input {...requiredProps} id="unique-id" ref={ref} />);
+      expect(ref.current.id).toBe('unique-id');
     });
   });
 
-  describe('invalid API', () => {
-    it('sets an invalid class', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          invalid
-        />
-      ));
-      expect(screen.getByLabelText('the label')).toHaveClass('invalid-input');
-    });
-
-    it('inserts an invalid icon', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          invalid
-        />
-      ));
-      expect(screen.getByRole('img')).toHaveClass('invalid-icon');
+  describe('label API', () => {
+    it('can be set', () => {
+      render(<Input {...requiredProps} label="Unique label" />);
+      expect(screen.getByLabelText('Unique label')).toBeInTheDocument();
     });
   });
 
   describe('maxLength API', () => {
     it('sets the maxLength attribute', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          maxLength={100}
-        />
-      ));
+      render(<Input {...requiredProps} maxLength={100} />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('maxlength', '100');
     });
   });
 
   describe('name API', () => {
     it('sets the name attribute', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          name="test-name"
-        />
-      ));
+      render(<Input {...requiredProps} name="test-name" />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('name', 'test-name');
+    });
+
+    it('is set on the ref', () => {
+      const ref = createRef();
+      render(<Input {...requiredProps} name="unique-name" ref={ref} />);
+      expect(ref.current.name).toBe('unique-name');
     });
   });
 
   describe('onBlur API', () => {
     it('sets the onblur handler', () => {
       const onBlurSpy = jest.fn();
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          onBlur={onBlurSpy}
-        />
-      ));
+      render(<Input {...requiredProps} onBlur={onBlurSpy} />);
       fireEvent.blur(screen.getByLabelText('the label'));
       expect(onBlurSpy.mock.calls.length).toEqual(1);
     });
@@ -155,13 +126,7 @@ describe('Input', () => {
   describe('onChange API', () => {
     it('sets the onchange handler', () => {
       const onChangeSpy = jest.fn();
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          onChange={onChangeSpy}
-        />
-      ));
+      render(<Input {...requiredProps} onChange={onChangeSpy} />);
       fireEvent.change(screen.getByLabelText('the label'), { target: { value: 'new value' } });
       expect(onChangeSpy.mock.calls.length).toEqual(1);
     });
@@ -170,13 +135,7 @@ describe('Input', () => {
   describe('onFocus API', () => {
     it('sets the onFocus handler', () => {
       const onFocusSpy = jest.fn();
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          onFocus={onFocusSpy}
-        />
-      ));
+      render(<Input {...requiredProps} onFocus={onFocusSpy} />);
       fireEvent.focus(screen.getByLabelText('the label'));
       expect(onFocusSpy.mock.calls.length).toEqual(1);
     });
@@ -185,13 +144,7 @@ describe('Input', () => {
   describe('onInput API', () => {
     it('sets the onInput handler', () => {
       const onInputSpy = jest.fn();
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          onInput={onInputSpy}
-        />
-      ));
+      render(<Input {...requiredProps} onInput={onInputSpy} />);
       fireEvent.input(screen.getByLabelText('the label'));
       expect(onInputSpy.mock.calls.length).toEqual(1);
     });
@@ -200,13 +153,7 @@ describe('Input', () => {
   describe('onKeyDown API', () => {
     it('sets the onKeyDown handler', () => {
       const onKeyDownSpy = jest.fn();
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          onKeyDown={onKeyDownSpy}
-        />
-      ));
+      render(<Input {...requiredProps} onKeyDown={onKeyDownSpy} />);
       fireEvent.keyDown(screen.getByLabelText('the label'));
       expect(onKeyDownSpy.mock.calls.length).toEqual(1);
     });
@@ -214,118 +161,120 @@ describe('Input', () => {
 
   describe('placeholder API', () => {
     it('sets the placeholder attribute', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          placeholder="test-placeholder"
-        />
-      ));
+      render(<Input {...requiredProps} placeholder="test-placeholder" />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('placeholder', 'test-placeholder');
     });
   });
 
   describe('readOnly API', () => {
     it('sets the readOnly attribute', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          readOnly
-        />
-      ));
+      render(<Input {...requiredProps} readOnly />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('readonly');
     });
   });
 
+  describe('ref API', () => {
+    it('can be set', () => {
+      const ref = React.createRef();
+      render(<Input {...requiredProps} ref={ref} value="test value" />);
+      expect(ref.current.value).toEqual('test value');
+    });
+  });
+
   describe('required API', () => {
-    it('sets the required attribute', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          required
-        />
-      ));
+    it('can be set', () => {
+      render(<Input {...requiredProps} required={true} />);
       expect(screen.getByLabelText('the label')).toBeRequired();
+      expect(screen.getByLabelText('the label')).toBeInvalid();
+    });
+
+    it('can be unset', () => {
+      render(<Input {...requiredProps} required={false} />);
+      expect(screen.getByLabelText('the label')).not.toBeRequired();
+      expect(screen.getByLabelText('the label')).toBeValid();
+    });
+
+    it('is invalid on mount but does not have an error message', () => {
+      render(<Input {...requiredProps} required={true} />);
+      expect(screen.getByLabelText('the label')).toBeInvalid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('');
+    });
+
+    it('is invalid after typing', async () => {
+      render(<Input {...requiredProps} required={true} />);
+      userEvent.type(screen.getByLabelText('the label'), 'a');
+      expect(screen.getByLabelText('the label')).toHaveValue('a');
+      expect(screen.getByLabelText('the label')).toBeValid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('');
+      userEvent.type(screen.getByLabelText('the label'), '{backspace}');
+      expect(screen.getByLabelText('the label')).toHaveValue('');
+      expect(screen.getByLabelText('the label')).toBeInvalid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('Constraints not satisfied');
+    });
+
+    it('is invalid after tabbing through', async () => {
+      render(<Input {...requiredProps} required={true} />);
+      userEvent.tab();
+      expect(document.activeElement).toBe(screen.getByLabelText('the label'));
+      expect(screen.getByLabelText('the label')).toBeInvalid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('');
+      userEvent.tab();
+      expect(document.activeElement).not.toBe(screen.getByLabelText('the label'));
+      expect(screen.getByLabelText('the label')).toBeInvalid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('Constraints not satisfied');
     });
   });
 
   describe('type API', () => {
     it('is set to "email"', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          type="email"
-        />
-      ));
+      render(<Input {...requiredProps} type="email" />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('type', 'email');
     });
 
     it('is set to "password"', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          type="password"
-        />
-      ));
+      render(<Input {...requiredProps} type="password" />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('type', 'password');
     });
 
     it('is be set to "text"', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          type="text"
-        />
-      ));
+      render(<Input {...requiredProps} type="text" />);
       expect(screen.getByLabelText('the label')).toHaveAttribute('type', 'text');
+    });
+  });
+
+  describe('validationMessage API', () => {
+    it('can be set', () => {
+      render(<Input {...requiredProps} validationMessage="unique error" />);
+      expect(screen.getByLabelText('the label')).toBeInvalid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('unique error');
+      expect(screen.getByRole('img', { name: 'unique error' })).toBeInTheDocument();
+    });
+
+    it('can be unset', () => {
+      render(<Input {...requiredProps} validationMessage="" />);
+      expect(screen.getByLabelText('the label')).toBeValid();
+      expect(screen.getByLabelText('the label')).toHaveDescription('');
     });
   });
 
   describe('value API', () => {
     it('sets the value', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          onChange={() => {}}
-          value="test-value"
-        />
-      ));
+      render(<Input {...requiredProps} value="test-value" />);
       expect(screen.getByLabelText('the label')).toHaveValue('test-value');
     });
-  });
 
-  describe('ref API', () => {
-    it('is set to the inputRef property', () => {
-      const inputRef = React.createRef();
-      render(
-        <Input
-          id="foo"
-          label="the label"
-          inputRef={inputRef}
-          onChange={() => {}}
-          value="test-value"
-        />,
-      );
-      expect(inputRef.current.value).toEqual('test-value');
+    it('updates the value', () => {
+      const { rerender } = render(<Input {...requiredProps} value="test value" />);
+      rerender(<Input {...requiredProps} value="another value" />);
+      expect(screen.getByLabelText('the label')).toHaveValue('another value');
     });
-  });
 
-  describe('autoFocus API', () => {
-    it('sets the autoFocus attribute', () => {
-      render((
-        <Input
-          id="foo"
-          label="the label"
-          autoFocus
-        />
-      ));
-      expect(screen.getByLabelText('the label')).toHaveFocus();
+    it('set on the ref', () => {
+      const ref = createRef();
+      render(<Input {...requiredProps} ref={ref} value="unique value" />);
+      expect(ref.current.value).toBe('unique value');
+      userEvent.type(screen.getByLabelText('the label'), '!');
+      expect(ref.current.value).toBe('unique value!');
     });
   });
 });