+++ CHANGELOG.md
@@ -7,7 +7,12 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
-
+  - Major: Deprecate unused props API on `CustomFieldInputText`: `ariaProps.autocomplete`, `ariaProps.haspopup`, `disabled`, `onBlur`, `onChange`, `onClick`, `onKeyDown`, `onKeyUp`
+  - Major: Deprecate `defaultValue` prop API on `CustomFieldInputText` in favor of `value`
+  - Major: Deprecate `disabled` prop API on `CustomFieldInputCurrency`
+  - Minor: Enhance `Input` component with validation logic, ref API
+  - Patch: Fix a11y descriptions for `Input` component
+  - Major: Deprecate unused props API `Input`: `invalid`, `defaultValue`
 </details>
 
 ## v0.49.0